deep_q_learning/requirements.txt
absl-py==0.9.0
astor==0.8.1
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cloudpickle==1.3.0
future==0.18.2
gast==0.2.2
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
gym==0.17.2
gym-retro==0.8.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-rl==0.4.2
Markdown==3.2.2
mock==4.0.2
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rl==3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==1.13.1
tensorboard-plugin-wit==1.7.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
